+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Release TBD
 - Make SPAMX optional, may interfer with form auto-complete.
 - Add option to record real IP addresses or anonymized addresses.
 - Use `ADMIN_list()` to show results table for more flexibility.
+- Implement moderation and notify other plugins of approval.
 
 ## v0.4.3
 Release 2019-05-22